refactor(owner-view): converge service/cause/research/event/ai_assistant onto public layout (SSOT)
Slice 2 of the owner-view refactor, following the product template (#339). Every
entity type that has a rich public detail page now renders that SAME layout for its
owner (dashboard) too, plus the manage bar — no more flat label→value grid.

SSOT / DRY / SoC:
- New src/components/public/detail-configs/{service,cause,research,event,ai-assistant}.tsx
  each export their EntityDetailConfig (moved verbatim from the public route, helpers
  + metadata logic preserved — pure move, zero behavior change).
- detail-configs/index.ts: barrel + DETAIL_CONFIGS registry (entityType → config) for
  generic lookups. One definition per entity, imported by both routes.
- Public routes (/services, /causes, /research, /events, /ai-assistants) slimmed to
  import their config.
- Dashboard routes (/dashboard/{services,causes,research,events,ai-assistants}/[id])
  now render PublicEntityDetailPage with the shared config — flat EntityDetailPage +
  makeDetailFields gone for these types.

Flat-layout types (assets/investments/loans/circles/groups/wishlists/documents) still
use EntityDetailPage for their public pages — they need their own configs, tracked as
the next chunk.

tsc 0; lint clean.
