+++ src/components/public/detail-configs/index.ts
@@ -1,0 +1,34 @@
+/**
+ * Entity detail-page configs (SSOT).
+ *
+ * Each entity's PublicEntityDetailPage config is defined ONCE here and imported by
+ * both its public route (/{type}/[id]) and its owner dashboard route
+ * (/dashboard/{base}/[id]) — same layout for buyers and owners, owner additionally
+ * gets the manage bar. DETAIL_CONFIGS maps entityType → config for generic lookups.
+ */
+import type { EntityType } from '@/config/entity-registry';
+import type { EntityDetailConfig } from '@/components/public/PublicEntityDetailPage';
+import { productDetailConfig } from './product';
+import { serviceDetailConfig } from './service';
+import { causeDetailConfig } from './cause';
+import { researchDetailConfig } from './research';
+import { eventDetailConfig } from './event';
+import { aiAssistantDetailConfig } from './ai-assistant';
+
+export {
+  productDetailConfig,
+  serviceDetailConfig,
+  causeDetailConfig,
+  researchDetailConfig,
+  eventDetailConfig,
+  aiAssistantDetailConfig,
+};
+
+export const DETAIL_CONFIGS: Partial<Record<EntityType, EntityDetailConfig>> = {
+  product: productDetailConfig,
+  service: serviceDetailConfig,
+  cause: causeDetailConfig,
+  research: researchDetailConfig,
+  event: eventDetailConfig,
+  ai_assistant: aiAssistantDetailConfig,
+};
